Chapter 6: Admin Controllers and Hooks

+++ mymodcomments.php
@@ -27,9 +27,15 @@ class MyModComments extends Module
 		if (!$this->loadSQLFile($sql_file))
 			return false;
 
+		// Install admin tab
+		if (!$this->installTab('AdminCatalog', 'AdminMyModComments', 'MyMod Comments'))
+			return false;
+
 		// Register hooks
 		if (!$this->registerHook('displayProductTabContent') ||
 			!$this->registerHook('displayBackOfficeHeader') ||
+			!$this->registerHook('displayAdminProductsExtra') ||
+			!$this->registerHook('displayAdminCustomers') ||
 			!$this->registerHook('ModuleRoutes'))
 			return false;
 
@@ -51,6 +57,10 @@ class MyModComments extends Module
 		// $sql_file = dirname(__FILE__).'/install/uninstall.sql';
 		// if (!$this->loadSQLFile($sql_file))
 		//	return false;
+
+		// Uninstall admin tab
+		if (!$this->uninstallTab('AdminMyModComments'))
+			return false;
 
 		// Delete configuration values
 		Configuration::deleteByName('MYMOD_GRADES');
@@ -79,6 +89,32 @@ class MyModComments extends Module
 		return $result;
 	}
 
+	public function installTab($parent, $class_name, $name)
+	{
+		// Create new admin tab
+		$tab = new Tab();
+		$tab->id_parent = (int)Tab::getIdFromClassName($parent);
+		$tab->name = array();
+		foreach (Language::getLanguages(true) as $lang)
+			$tab->name[$lang['id_lang']] = $name;
+		$tab->class_name = $class_name;
+		$tab->module = $this->name;
+		$tab->active = 1;
+		return $tab->add();
+	}
+
+	public function uninstallTab($class_name)
+	{
+		// Retrieve Tab ID
+		$id_tab = (int)Tab::getIdFromClassName($class_name);
+
+		// Load tab
+		$tab = new Tab((int)$id_tab);
+
+		// Delete it
+		return $tab->delete();
+	}
+
 	public function onClickOption($type, $href = false)
 	{
 		$confirm_reset = $this->l('Reseting this module will delete all comments from your database, are you sure you want to reset it ?');
@@ -103,7 +139,7 @@ class MyModComments extends Module
 		// Build dynamically the controller name
 		$controller_name = $this->name.$hook_name.'Controller';
 
-		// Instanciate controller
+		// Instantiate controller
 		$controller = new $controller_name($this, __FILE__, $this->_path);
 
 		// Return the controller
@@ -122,6 +158,18 @@ class MyModComments extends Module
 		return $controller->run($params);
 	}
 
+	public function hookDisplayAdminProductsExtra($params)
+	{
+		$controller = $this->getHookController('displayAdminProductsExtra');
+		return $controller->run();
+	}
+
+	public function hookDisplayAdminCustomers($params)
+	{
+		$controller = $this->getHookController('displayAdminCustomers');
+		return $controller->run();
+	}
+
 	public function hookModuleRoutes()
 	{
 		$controller = $this->getHookController('modulesRoutes');
@@ -130,6 +178,14 @@ class MyModComments extends Module
 
 	public function getContent()
 	{
+		$ajax_hook = Tools::getValue('ajax_hook');
+		if ($ajax_hook != '')
+		{
+			$ajax_method = 'hook'.ucfirst($ajax_hook);
+			if (method_exists($this, $ajax_method))
+				die($this->{$ajax_method}(array()));
+		}
+
 		$controller = $this->getHookController('getContent');
 		return $controller->run();
 	}
